requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0

# Report Generation
reportlab>=4.0.0
openpyxl>=3.1.0
pandas>=2.0.0

# Backend API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
fastapi[standard]

# Frontend UI
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging and debugging
rich>=13.7.0

# Data handling
requests>=2.31.0

